+++ ZoranEngine/Core/CameraBase.h
@@ -13,11 +13,11 @@ public:
 
 	virtual MatrixF GetModel() = 0;
 
-	virtual void Translate(double deltax, double deltay, double deltaz) = 0;
+	virtual void Translate(float deltax, float deltay, float deltaz) = 0;
 	virtual void Translate(Vec2D delta) = 0;
 	virtual void Translate(Vec3D delta) = 0;
-	virtual void Rotate(double rotx, double roty, double rotz) = 0;
-	virtual void Rotate(double rotation) = 0;
+	virtual void Rotate(float rotx, float roty, float rotz) = 0;
+	virtual void Rotate(float rotation) = 0;
 	virtual void Rotate(Vec3D euler) = 0;
 	virtual void Rotate(Quaternion quat) = 0;
 };
